+++ src/OneBlinkFormBase.tsx
@@ -40,6 +40,7 @@ import checkBsbsAreInvalid from './services/checkBsbsAreInvalid'
 import checkIfBsbsAreValidating from './services/checkIfBsbsAreValidating'
 import checkIfAttachmentsExist from './services/checkIfAttachmentsExist'
 import useAuth from './hooks/useAuth'
+import determineIsInfoPage from './services/determineIsInfoPage'
 
 export type BaseProps = {
   onCancel: () => unknown
@@ -56,6 +57,8 @@ export type BaseProps = {
   onSaveDraft?: (
     newDraftSubmission: submissionService.NewDraftSubmission,
   ) => unknown
+  handleNavigateAway?: () => unknown
+  isInfoPage?: 'YES' | 'NO' | 'CALCULATED'
 }
 
 export type ControlledProps = {
@@ -86,6 +89,8 @@ function OneBlinkFormBase({
   primaryColour,
   attachmentRetentionInDays,
   allowSubmitWithPendingAttachments,
+  handleNavigateAway,
+  isInfoPage: isInfoPageProp,
 }: Props) {
   const isOffline = useIsOffline()
   const { isUsingFormsKey } = useAuth()
@@ -104,6 +109,13 @@ function OneBlinkFormBase({
       }),
     [primaryColour],
   )
+
+  const isInfoPage = React.useMemo(() => {
+    if (!!isInfoPageProp && isInfoPageProp !== 'CALCULATED') {
+      return isInfoPageProp === 'YES'
+    }
+    return determineIsInfoPage(definition)
+  }, [definition, isInfoPageProp])
 
   //
   //
@@ -198,11 +210,20 @@ function OneBlinkFormBase({
       // Navigate to the previous blocked location with your navigate function
       if (goToLocation) {
         history.push(`${goToLocation.pathname}${goToLocation.search}`)
+        if (handleNavigateAway) {
+          handleNavigateAway()
+        }
       } else {
         onCancel()
       }
     }
-  }, [goToLocation, hasConfirmedNavigation, history, onCancel])
+  }, [
+    goToLocation,
+    handleNavigateAway,
+    hasConfirmedNavigation,
+    history,
+    onCancel,
+  ])
 
   const handleCancel = React.useCallback(() => {
     if (isDirty) {
@@ -560,7 +581,16 @@ function OneBlinkFormBase({
 
   const handleChange = React.useCallback<FormElementValueChangeHandler>(
     (element, value) => {
-      if (disabled || element.type === 'page' || element.type === 'section') {
+      if (
+        //This will ensure on a read only form that the summary and calculation elements
+        //can still be displayed as it needs handleChange so it can render
+        //due to the dynamic nature of the summary element.
+        (disabled &&
+          element.type !== 'summary' &&
+          element.type !== 'calculation') ||
+        element.type === 'page' ||
+        element.type === 'section'
+      ) {
         return
       }
 
@@ -843,7 +873,7 @@ function OneBlinkFormBase({
             </div>
             {!isReadOnly && (
               <div className="buttons ob-buttons ob-buttons-submit">
-                {onSaveDraft && !definition.isInfoPage && (
+                {onSaveDraft && !isInfoPage && (
                   <button
                     type="button"
                     className="button ob-button is-primary ob-button-save-draft cypress-save-draft-form"
@@ -857,7 +887,7 @@ function OneBlinkFormBase({
                   </button>
                 )}
                 <span className="ob-buttons-submit__spacer"></span>
-                {!definition.isInfoPage && (
+                {!isInfoPage && (
                   <button
                     type="button"
                     className="button ob-button is-light ob-button-submit-cancel cypress-cancel-form"
@@ -878,9 +908,7 @@ function OneBlinkFormBase({
                   >
                     <CustomisableButtonInner
                       label={
-                        definition.isInfoPage
-                          ? 'Done'
-                          : buttons?.submit?.label || 'Submit'
+                        isInfoPage ? 'Done' : buttons?.submit?.label || 'Submit'
                       }
                       icon={buttons?.submit?.icon}
                     />