+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Data Manager query string management to merge instead of replace
 
+### Fixed
+
+- Autosave form not discarding autosave
+
 ## [2.5.2] - 2023-01-16
 
 ### Fixed