+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Summary Element not showing on read only forms
+- Summary and Calculation Elements not showing on read only forms
 
 ## [2.5.2] - 2023-01-16