+++ CHANGELOG.md
@@ -7,8 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+`isInfoPage` prop
+
+### Changed
+
+- Data Manager query string management to merge instead of replace
+
 ### Fixed
 
+- Autosave form not discarding autosave
+- Summary and Calculation Elements not showing on read only forms
 - number input not allowing zeros
 
 ## [2.5.2] - 2023-01-16